ongoing work - filters refactoring, added FilterStateComponent service
