+++ src/app/ontology-improver/ontology-improver.component.ts
@@ -10,6 +10,7 @@ import {ActivatedRoute, Params, Router} from '@angular/router';
 import {Title} from '@angular/platform-browser';
 import {FormGroup, Validators, FormBuilder} from '@angular/forms';
 import {ApiDataService} from '../services/api-data.service';
+import {FilterStateService} from '../services/filter-state.service';
 import {validation_ws_url} from '../shared/constants';
 
 @Component({
@@ -92,6 +93,7 @@ export class OntologyImproverComponent implements OnInit, OnDestroy {
 
   constructor(
     private dataService: ApiDataService,
+    private filterStateService: FilterStateService,
     private ontologyService: OntologyService,
     public dialog: MatDialog,
     public snackbar: MatSnackBar,
@@ -140,7 +142,7 @@ export class OntologyImproverComponent implements OnInit, OnDestroy {
       this.aggregationService.getAggregations(data.aggregations, 'ontology');
     });
 
-    this.updateUrlParams();
+    this.aggrSubscription = this.filterStateService.updateUrlParams(this.query, ['ontology']);
 
     // fetch usage statistics summary
     this.ontologyService.getUsageStatistics().subscribe((data) => {
@@ -705,24 +707,7 @@ export class OntologyImproverComponent implements OnInit, OnDestroy {
   }
 
   loadInitialPageState(params){
-    const filters = {};
-    // set up filters on pageLoad based on queryParams
-    for (const key in params) {
-      if (key !== 'searchTerm' && key !== 'sortTerm' && key !== 'sortDirection' && key !== 'pageIndex'){
-        if (Array.isArray(params[key])) {
-          filters[key] = params[key];
-          for (const value of params[key]) {
-            this.aggregationService.current_active_filters.push(value);
-            this.aggregationService.active_filters[key].push(value);
-          }
-        } else {
-          filters[key] = [params[key]];
-          this.aggregationService.current_active_filters.push(params[key]);
-          this.aggregationService.active_filters[key].push(params[key]);
-        }
-      }
-    }
-    this.aggregationService.field.next(this.aggregationService.active_filters);
+    const filters = this.filterStateService.setUpAggregationFilters(params);
     this.filter_field = filters;
     this.query['filters'] = filters;
     // load pre-search and pre-sorting
@@ -733,29 +718,6 @@ export class OntologyImproverComponent implements OnInit, OnDestroy {
       this.query['sort'] = [params['sortTerm'], params['sortDirection']];
     }
   }
-
-
-  updateUrlParams(){
-    // setting urls params based on filters
-    this.aggrSubscription = this.aggregationService.field.subscribe((data) => {
-      const params = {};
-      for (const key of Object.keys(data)) {
-        if (data[key] && data[key].length !== 0) {
-          params[key] = data[key];
-        }
-      }
-      //update url for search term and sorting
-      if (this.query['search']){
-        params['searchTerm'] = this.query['search'];
-      }
-      if (this.query['sort']){
-        params['sortTerm'] = this.query['sort'][0]
-        params['sortDirection'] = this.query['sort'][1]
-      }
-      this.router.navigate(['ontology'], {queryParams: params});
-    });
-  }
-
 
   ngOnDestroy() {
     if (typeof this.filter_field !== 'undefined') {
